requirements.txt
webdriver_manager
selenium
aiohttp
beautifulsoup4
rich
google-generativeai
jinja2
